tests/wpt/tests/css/css-gaps/tentative/parsing/gap-decorations-width-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Gap Decorations: column-rule-width parsing</title>
<link rel="help" href="https://kbabbitt.github.io/css-gap-decorations/#column-row-rule-width">
<link rel="author" title="Sam Davis Omekara Jr." href="mailto:samomekarajr@microsoft.com">
<meta name="assert" content="column-rule-width supports only the grammar '[ <line-width-list> | <auto-line-width-list> ]'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("column-rule-width", "auto");

test_invalid_value("column-rule-width", "thin, medium, thick");
test_invalid_value("column-rule-width", "repeat(auto, thin, 10px, thick)");
test_invalid_value("column-rule-width", "repeat(0, 10px, 20px, 30px)");
test_invalid_value("column-rule-width", "repeat(-1, 10px, thin, medium)");
test_invalid_value("column-rule-width", "repeat(auto, 10px) thin repeat(auto, thick)");
test_invalid_value("column-rule-width", "10 20 30");
test_invalid_value("column-rule-width", "repeat(2, -20px)");
test_invalid_value("column-rule-width", "30%");
</script>
</body>
</html>
